식단 캘린더 frontend  수정

+++ backend/routes/users.js
@@ -266,13 +266,14 @@ router.post(
         const userInfo = await users.findOne({
           where: { id: req.params.id },
         });
-
         const s3key = userInfo.s3_key;
-        const deleteParams = {
-          Bucket: "fitme-s3",
-          Key: s3key,
-        };
-        await s3.deleteObject(deleteParams).promise();
+        if (s3key != null) {
+          const deleteParams = {
+            Bucket: "fitme-s3",
+            Key: s3key,
+          };
+          await s3.deleteObject(deleteParams).promise();
+        }
         const uploadParams = {
           acl: "public-read",
           ContentType: "image/png",
